separated file loading to json_loader.js from json_parser.js
